+++ app/helpers/lookups_helper.rb
@@ -1,0 +1,2 @@
+module LookupsHelper
+end