+++ config/application.rb
@@ -10,11 +10,10 @@ require 'action_mailer/railtie'
 Bundler.require(:default, Rails.env) if defined?(Bundler)
 
 module Snorby
-  
   # Snorby Environment Specific Configurations
   raw_config = File.read("config/snorby_config.yml")
   CONFIG = YAML.load(raw_config)[Rails.env].symbolize_keys
-  
+
   class Application < Rails::Application
         
     PDFKit.configure do |config|
@@ -31,7 +30,7 @@ module Snorby
 
     # Custom directories with classes and modules you want to be autoloadable.
     config.autoload_paths += %W(#{config.root}/lib)
-
+    
     # Only load the plugins named here, in the order given (default is alphabetical).
     # :all can be used as a placeholder for all plugins not explicitly named.
     # config.plugins = [ :exception_notification, :ssl_requirement, :all ]