+++ public/javascripts/snorby.js
@@ -468,7 +468,33 @@ var Snorby = {
 				$('pre#' + $(this).attr('data-content-name')).show();
 				return false;
 			});
-			
+		
+      $('a.snorbybox-content').live('click', function(event) {
+        event.preventDefault();
+        $('dl.drop-down-menu').fadeOut('slow');
+				var content = $(this).attr('data-content');
+
+        $.fancybox({
+					padding: 0,
+          content: content,
+					centerOnScroll: true,
+	        zoomSpeedIn: 300, 
+	        zoomSpeedOut: 300,
+					overlayShow: true,
+					overlayOpacity: 0.5,
+					overlayColor: '#000',
+					onStart: function() {
+						$(document).unbind('keydown', 'right');
+						$(document).unbind('keydown', 'esc');
+						$(document).unbind('keydown', 'shift+left');
+						$('dl#event-sub-menu').hide();
+					},
+					onClosed: function() {
+						Snorby.hotkeys();
+					}
+				});
+      });
+
 			$('a.snorbybox').live('click', function() {
 				$('dl.drop-down-menu').fadeOut('slow');
 				$.fancybox({