+++ spec/models/lookup_spec.rb
@@ -1,0 +1,5 @@
+require 'spec_helper'
+
+describe Lookup do
+  pending "add some examples to (or delete) #{__FILE__}"
+end